Move hush.conf parsing code into a module (hush.Config).
