+++ lib/hush/package.json
@@ -1,0 +1,7 @@
+{
+    "name": "hush",
+    "description": "Hush helpers",
+    "private": true,
+    "version": "0.0.1",
+    "main": "index.js"
+  }
