+++ package.json
@@ -54,6 +54,7 @@
     "axios": "^0.16.1",
     "bitcoin": "^3.0.1",
     "copy-to-clipboard": "^3.0.8",
+    "hush": "file:lib/hush",
     "node-cmd": "^3.0.0",
     "repeat": "^0.0.6",
     "request": "^2.83.0",
